practices/ImageVectorSearch/requirements.txt
wget
azure-search-documents
azure-identity
pandas
openai
